Refactor: creates profile module.
