+++ src/app/product/components/index.js
@@ -1,0 +1,5 @@
+export {default as BoardSection} from './BoardSection'
+export {default as ReplyModal} from './ReplyModal'
+export {default as ReviewModal} from './ReviewModal'
+export {default as ReviewsSection} from './ReviewsSection'
+export {default as ResourcesSection} from './ResourcesSection'
