+++ src/app/profile/components/profile/index.js
@@ -1,0 +1,3 @@
+export {default as EditProfileModal} from './EditProfileModal'
+export {default as ScoreModal} from './ScoreModal'
+export {default as ProfileInfo} from './ProfileInfo'
